requirements.txt
Flask==2.3.3
qrcode[pil]
fpdf
Pillow
PyDrive
gunicorn
cloudinary
